+++ test/test_helper.exs
@@ -1,7 +1,6 @@
 e2e? = System.get_env("GPUI_E2E") == "1"
 
 Code.require_file("../support/ssl_certs.exs", __DIR__)
-Code.require_file("../support/test_display.ex", __DIR__)
 
 if e2e? do
   Code.require_file("support/e2e/desktop.ex", __DIR__)